feat: Implement RegisterController for user registration
