+++ src/main/java/com/f5/buzon_inteligente_BE/credencial/CredencialController.java
@@ -1,27 +1,29 @@
-package com.f5.buzon_inteligente_BE.user;
-
-import com.f5.buzon_inteligente_BE.dto.CredentialRequest;
+package com.f5.buzon_inteligente_BE.credencial;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 
-@RestController
-@RequestMapping("/api/credential")
-public class UserController {
 
+
+public class CredencialController {
+
+    
     @Autowired
-    private UserService userService;
+    private CredencialService credencialService;
 
     @PutMapping("/permanent/{userId}")
     public ResponseEntity<String> updateCredential(
             @PathVariable Long userId,
             @RequestBody CredentialRequest request) {
         try {
-            boolean updated = userService.updatePermanentCredential(userId, request);
+            boolean updated = credencialService.updatePermanentCredential(userId, request);
             return ResponseEntity.ok("Credential updated successfully.");
         } catch (RuntimeException e) {
             return ResponseEntity.badRequest().body(e.getMessage());
         }
     }
+
 }
