+++ src/main/java/com/f5/buzon_inteligente_BE/user/UserService.java
@@ -1,38 +1,29 @@
 package com.f5.buzon_inteligente_BE.user;
 
-import com.f5.buzon_inteligente_BE.dto.CredentialRequest;
-
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
 import java.util.Optional;
 
 @Service
 public class UserService {
 
-    @Autowired
-    private UserRepository userRepository;
+    private final UserRepository userRepository;
 
-    public boolean updatePermanentCredential(Long userId, CredentialRequest request) {
-        Optional<User> userOptional = userRepository.findById(userId);
-        if (!userOptional.isPresent()) {
-            throw new RuntimeException("User not found");
-        }
-        
-        User user = userOptional.get();
-        
-        if (!user.getUserPassword().equals(request.getCurrentPassword())) {
-            throw new RuntimeException("Incorrect password");
-        }
-        
-        Optional<User> existingCredential = userRepository.findByPermanentCredential(request.getNewCredential());
-        if (existingCredential.isPresent()) {
-            throw new RuntimeException("Credential is already in use");
-        }
-        
-        user.setPermanentCredential(request.getNewCredential());
-        userRepository.save(user);
-
-        return true; 
+   
+    public UserService(UserRepository userRepository) {
+        this.userRepository = userRepository;
     }
+
+    public Optional<User> findByEmail(String email) {
+        return userRepository.findByUserEmail(email);
+    }
+
+    public Optional<User> findById(Long id) {
+        return userRepository.findById(id);
+    }
+
+    public User save(User user) {
+        return userRepository.save(user);
+    }
+
+   
 }
